fix: onboarding phase failed to start for new user
The problem caused by mismatch version of raw_user_metadata, the component that decide which phase to render (RegistrationPageIndex) listen to session.user.user_metadata.registration_phase, the thing is when the server set the registration_phase flag to "name" for new user it only change the user_metadata field on DB and the client left with the old unchange version (no registration_phase flag).

The solution is to setup a default value, always assume the phase for a user when it was redirected to auth/registration is "name". By that, the UI render quickly. When the user updated the flag, the refetch will work perfectly fine. When the user leave the onboarding process before leaving it, the new session will already have the flag so no problem at all, in fact it's actually fix the temporary blank_page problem too